Saf and Dummy providers return properly different JWT tokens on successive authentication
