+++ gradle.properties
@@ -25,7 +25,7 @@ nodejsVersion=8.11.1
 projectRoot=${project.projectDir}
 
 org.gradle.daemon=false
-org.gradle.jvmargs=-Xmx1536m
+org.gradle.jvmargs=-Xmx4096m
 org.gradle.parallel=true
 org.gradle.caching=true
 # Don't push in remote cache from local environments
